requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pkg-resources==0.0.0
PyJWT==1.7.1
rope==0.17.0
Werkzeug==1.0.1
